+++ app/models/domain_file_version.rb
@@ -14,6 +14,17 @@ class DomainFileVersion < DomainModel
   belongs_to :creator, :class_name => 'EndUser',:foreign_key => 'creator_id'
   
   def self.archive(file)
+
+    if !file.processor_handler.revision_support
+      begin
+        file.destroy_thumbs
+        File.unlink(file.filename)
+      rescue Exception => e
+        raise e
+        # Chomp
+      end
+      return
+    end
   
     info = {}
     if file.image_size
@@ -21,6 +32,10 @@ class DomainFileVersion < DomainModel
     end
     
     version_hash = self.generate_version_hash
+
+    file.processor_handler.destroy_remote! # get rid of remote copies of this file
+
+    file.filename # copy file locally
     
     # Create a new instance with the correct file attributes
     v = DomainFileVersion.create(:domain_file => file,
@@ -32,13 +47,17 @@ class DomainFileVersion < DomainModel
                                  :file_size => file.file_size,
                                  :creator_id => file.creator_id,
                                  :version_hash => version_hash,
-                                 :stored_at => file.stored_at)
+                                 :stored_at => file.stored_at,
+                                 :server_id => Server.server_id)
   end
   
   
   # Create a domain file from this file
   def extract(user_id=nil)
-    df = DomainFile.new(:parent_id => self.domain_file.parent_id, :creator_id => user_id || self.creator_id, :private => self.domain_file.private)
+    self.copy_local!
+    return nil unless File.exists?(self.abs_filename)
+
+    df = DomainFile.new(:parent_id => self.domain_file.parent_id, :creator_id => user_id || self.creator_id, :private => self.domain_file.private, :process_immediately => true)
     File.open(self.abs_filename,"rb") do |f|
       df.filename = f
       if(df.save)
@@ -46,13 +65,24 @@ class DomainFileVersion < DomainModel
       end
     end
   end
-  
+
+  def extract_file(options={})
+    file = self.extract
+    {:domain_file_id => file.id} if file
+  end
+
+  def prefixed_filename
+    # unless we have a filename, return false
+    atr = self.read_attribute(:filename)
+    return nil unless self.prefix && atr
+     "#{DomainFile.storage_subdir}/#{self.prefix}/#{atr}"
+  end
   
    def relative_filename
       # unless we have a filename, return false
       atr = self.read_attribute(:filename)
       return nil unless self.prefix && atr
-      self.storage_directory + atr
+    self.storage_directory + atr
    end
    
    # Return the absolute storage directory - valid for opening a file  on the server 
@@ -70,14 +100,20 @@ class DomainFileVersion < DomainModel
    
    def storage_base; self.domain_file.private? ? DomainFile.private_storage_base : DomainFile.public_storage_base; end
    
-   def url; relative_filename; end
+   def url
+     if domain_file.processor != 'local'
+       self.domain_file.processor_handler.version_url(self)
+     else
+       relative_filename
+     end
+   end
    
    ###########
    # Hooks
    ###########
    
    def erase_version_file
-    self.domain_file.processor_handler.destroy_remote_version!(self) if self.domain_file.processor != 'local'
+    self.domain_file.processor_handler.destroy_remote_version!(self)
     if !prefix.blank? && (File.directory?(abs_storage_directory))
       FileUtils.rm_rf(abs_storage_directory)
     end
@@ -99,8 +135,16 @@ class DomainFileVersion < DomainModel
     return nil unless self.meta_info[:image_size]
     self.meta_info[:image_size][:height]
   end
-  
- protected 
+
+  def server
+    @server ||= Server.find_by_id(self.server_id) if self.server_id
+  end
+
+  def copy_local!
+    self.domain_file.processor_handler.respond_to?('copy_version_local!') ? self.domain_file.processor_handler.copy_version_local!(self) : false
+  end
+
+ protected
   def self.generate_version_hash
     now = Time.now
     digest  = Digest::SHA1.hexdigest("#{now}#{rand}#{now.usec}#{Process.pid}")[0..31]
